refactor(webtransportgo): drop func-fields from FakeStreamConn, rename file
No call site configured any of the callbacks — the double was only ever
used as a zero-value stub — so the func-fields collapse to value
overrides for address/TLS/close. Renames mock_stream_conn_test.go to
fake_stream_conn_test.go to match the fake_[feature]_test.go convention.
